version 1.15

+++ src/main/java/geneticalgorithm/GeneticAlgorithmGeneration.java
@@ -72,7 +72,7 @@ class GeneticAlgorithmGeneration {
         LOG.log(Level.INFO, () -> String.format("generation #%d: \t %s", id, populationList.get(0).getLogMessage()));
 
         if (populationList.get(0).isPerfectScore()) {
-            double scorePercent = 100 / populationSize * populationList.stream().filter(o -> o.isPerfectScore()).count();
+            double scorePercent = 100 / populationSize * (double) populationList.stream().filter(o -> o.isPerfectScore()).count();
             LOG.log(Level.INFO, () -> String.format("****************** PERFECT SCORE ACHIEVED! ****************** \nat generation #%d, %.2f%s units reached a perfect score.", id, scorePercent, "%"));
         }
